aecm: eliminate floating-point operations and ensure zero-initialization
Convert delay estimator histogram to Q14 fixed-point integer math and replace
floating-point scaling in AECm to achieve pure fixed-point execution with
zero soft-float symbols on embedded DSPs.

+++ webrtc/modules/audio_processing/utility/delay_estimator.h
@@ -51,8 +51,8 @@ typedef struct {
   int last_candidate_delay;
   int compare_delay;
   int candidate_hits;
-  float* histogram;
-  float last_delay_histogram;
+  int32_t* histogram;
+  int32_t last_delay_histogram;
 
   // For dynamically changing the lookahead when using SoftReset...().
   int lookahead;
